# Backend/src/team/team.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
} from '@nestjs/common';
import { TeamService } from './team.service';
import { CreateTeamDto } from './dto/create-team.dto';
import { UpdateTeamDto } from './dto/update-team.dto';

@Controller('teams')
export class TeamController {
  constructor(private readonly teamService: TeamService) {}

  @Post()
  create(@Body() createTeamDto: CreateTeamDto) {
    return this.teamService.create(createTeamDto);
  }

  @Post('/:teamID/addPlayer/:playerID')
  async addPlayerToTeam(
    @Param('teamID') teamID: string,
    @Param('playerID') playerID: string,
  ) {
    return this.teamService.addPlayerToTeam(+teamID, +playerID);
  }

  @Get()
  findAll() {
    return this.teamService.findAll(false);
  }

  @Get('players')
  findAllWithPlayers() {
    return this.teamService.findAll(true);
  }

  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.teamService.findOne(+id);
  }

  @Patch(':id')
  update(@Param('id') id: string, @Body() updateTeamDto: UpdateTeamDto) {
    return this.teamService.update(+id, updateTeamDto);
  }

  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.teamService.remove(+id);
  }
}